+++ antea/reco/petit_mc_reco_functions.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-import antea.reco.data_reco_functions as pf
+import antea.reco.petit_data_reco_functions as pf
 
 
 def is_a_coincidence(df: pd.DataFrame) -> bool:
