+++ antea/scripts/process_data_petit.py
@@ -2,7 +2,7 @@ import sys
 import numpy  as np
 import pandas as pd
 
-import antea.reco.data_reco_functions as drf
+import antea.reco.petit_data_reco_functions as drf
 
 
 def from_ToT_to_pes(x):
